Internal: refactor plot message path functions to separate file
